# contract/openapi.yaml
openapi: 3.1.0
info:
  title: AgentMux Public Integration API
  description: >
    Public integration contract for AgentMux. Stability tiers (stable / beta /
    internal) are defined in CONTRACT.md; only stable and beta endpoints are
    described here. The OpenAI-compatible surface (/v1/responses, /v1/files)
    follows the upstream OpenAI Responses specification and is intentionally
    not duplicated in this document.
  version: "1.3"
  x-contract-version: "1.3"
servers:
  - url: http://127.0.0.1:8765
    description: Local AgentMux daemon (default listen address)
security:
  - bridgeToken: []
...
paths:
...
  /api/v1/agent-instances:
    get:
      tags: [agents]
      operationId: listAgentInstances
      summary: List Agent instances
      responses:
        "200":
          description: Agent instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentInstance"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [agents]
      operationId: upsertAgentInstance
      summary: Create or update an Agent instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInstance"
      responses:
        "200":
          description: Stored Agent instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInstance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [agents]
      operationId: deleteAgentInstance
      summary: Delete an Agent instance
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bridgeToken:
      type: http
      scheme: bearer
      description: >
        The [bridge] token from config.toml. Required on /api/* and /v1/*
        whenever bridge.enabled = true. A valid agentmux_console session
        cookie is an equivalent credential for browser-embedded Console use.
  responses:
    OK:
      description: Success
      content:
        application/json:
          schema:
            type: object
            required: [ok]
            properties:
              ok: { type: boolean, const: true }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid bridge token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AgentInstance:
      type: object
      required: [id, name, runtime_id, enabled]
      properties:
        id: { type: string }
        name: { type: string }
        runtime_id: { type: string }
        desktop_thread_id: { type: string }
        work_dir: { type: string }
        workspace_mode: { type: string }
        worktree_base_ref: { type: string }
        session_backend: { type: string }
        system_prompt: { type: string }
        provider_tool: { type: string }
        provider_id: { type: string }
        provider_name: { type: string }
        default_model: { type: string }
        default_reasoning_effort: { type: string }
        default_service_tier: { type: string }
        default_approval_mode: { type: string }
        memory_scope: { type: string }
        env:
          type: object
          additionalProperties: { type: string }
        channel_bindings:
          type: array
          items:
            type: object
            additionalProperties: true
        schedules:
          type: array
          items:
            type: object
            additionalProperties: true
        mcp_servers:
          type: array
          items: { type: string }
        skills:
          type: array
          items: { type: string }
        clis:
          type: array
          items: { type: string }
        enabled: { type: boolean }
        source: { type: string }
        owner_tenant_id:
          type: string
          description: >
            Owning tenant, empty when the resource is unassigned. A tenant
            credential cannot set this: it is always stamped with the caller.
        owner_tenant_name:
          type: string
          description: Owner display name, filled on admin reads only.
        visibility:
          type: string
          enum: [private, public]
          description: >
            public makes the resource readable and runnable by every tenant;
            only its owner or the administrator may modify it.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...