# server/agent_instances.go
package server

import (
	"context"
	"crypto/rand"
	"encoding/hex"
	"encoding/json"
	"fmt"
	"io"
	"net/http"
	"sort"
	"strconv"
	"strings"
	"time"

	"github.com/wangning19940904/AgentMux/core"
	"github.com/wangning19940904/AgentMux/framework"
	sessionstore "github.com/wangning19940904/AgentMux/sessions"
)

func (s *Server) handleAgentInstancesList(w http.ResponseWriter, r *http.Request) {
	items, err := s.scopedAgentInstances(r.Context(), requestPrincipal(r))
	if err != nil {
		writeErr(w, http.StatusInternalServerError, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, items)
}

func (s *Server) handleAgentInstanceUpsert(w http.ResponseWriter, r *http.Request) {
	if s.st == nil {
		writeErr(w, http.StatusServiceUnavailable, "store unavailable")
		return
	}
	a, ok := decodeJSON[core.AgentInstance](w, r)
	if !ok {
		return
	}
	principal := requestPrincipal(r)
	if providerID := strings.TrimSpace(a.ProviderID); principal.IsTenant() && providerID != "" {
		if _, authorized := s.authorizeProvider(w, r, providerID, core.GrantLevelUse); !authorized {
			return
		}
	}
	// Editing an existing agent requires manage access on it. An unknown ID is
	// a create with a caller-chosen ID, which needs no prior authorization.
	var existing *core.AgentInstance
	if id := strings.TrimSpace(a.ID); id != "" {
		found, err := s.st.GetAgentInstance(r.Context(), id)
		if err != nil {
			writeErr(w, http.StatusInternalServerError, err.Error())
			return
		}
		if found != nil {
			if _, authorized := s.authorizeAgent(w, r, id, core.GrantLevelManage); !authorized {
				return
			}
			existing = found
		}
	}
	if err := s.normalizeAgentInstance(r.Context(), &a); err != nil {
		writeErr(w, http.StatusBadRequest, err.Error())
		return
	}
	s.stampAgentOwnership(r.Context(), principal, &a, existing)
	if err := s.st.UpsertAgentInstance(r.Context(), &a); err != nil {
		writeErr(w, http.StatusInternalServerError, err.Error())
		return
	}
	if !principal.IsTenant() && a.ProviderID != "" && a.ProviderTool != "" && s.provider != nil {
		if err := s.provider.Switch(r.Context(), a.ProviderID, a.ProviderTool); err != nil {
			writeErr(w, http.StatusInternalServerError, "agent saved, but provider route failed: "+err.Error())
			return
		}
	}
	if s.connect != nil {
		if err := s.connect.RestartChannelsForAgent(r.Context(), a.ID); err != nil {
			writeErr(w, http.StatusInternalServerError, "agent saved, but bound channels failed to restart: "+err.Error())
			return
		}
	}
	items := []core.AgentInstance{a}
	s.enrichAgentProviders(r.Context(), items, principal)
	s.labelAgentOwners(r.Context(), items)
	writeJSON(w, http.StatusOK, &items[0])
}

func (s *Server) handleAgentInstanceInitialize(w http.ResponseWriter, r *http.Request) {
	if s.workspace == nil {
		writeErr(w, http.StatusServiceUnavailable, "workspace initializer unavailable")
		return
	}
	var opts core.WorkspaceInitOptions
	if err := json.NewDecoder(r.Body).Decode(&opts); err != nil && err != io.EOF {
		writeErr(w, http.StatusBadRequest, err.Error())
		return
	}
	opts.AgentID = strings.TrimSpace(opts.AgentID)
	opts.RuntimeID = strings.TrimSpace(opts.RuntimeID)
	opts.WorkDir = strings.TrimSpace(opts.WorkDir)
	if opts.AgentID == "" {
		opts.AgentID = strings.TrimSpace(r.URL.Query().Get("id"))
	}
	if opts.AgentID != "" {
		inst, ok, err := s.findAgentInstance(r.Context(), opts.AgentID)
		if err != nil {
			writeErr(w, http.StatusInternalServerError, err.Error())
			return
		}
		if !ok {
			writeErr(w, http.StatusNotFound, "agent not found")
			return
		}
		if opts.RuntimeID == "" {
			opts.RuntimeID = inst.RuntimeID
		}
		if opts.WorkDir == "" {
			opts.WorkDir = inst.WorkDir
		}
		if opts.WorkspaceMode == "" {
			opts.WorkspaceMode = inst.WorkspaceMode
		}
		if opts.WorktreeBaseRef == "" {
			opts.WorktreeBaseRef = inst.WorktreeBaseRef
		}
		if len(opts.Skills) == 0 {
			opts.Skills = append([]string(nil), inst.Skills...)
		}
		if len(opts.MCPServers) == 0 {
			opts.MCPServers = append([]string(nil), inst.MCPServers...)
		}
	}
	res, err := s.workspace.InitializeWorkspace(r.Context(), opts)
	if err != nil {
		writeErr(w, http.StatusInternalServerError, err.Error())
		return
	}
	writeJSON(w, http.StatusOK, res)
}

func (s *Server) handleAgentInstanceDelete(w http.ResponseWriter, r *http.Request) {
	if s.st == nil {
		writeErr(w, http.StatusServiceUnavailable, "store unavailable")
		return
	}
	id, ok := requireQuery(w, r, "id")
	if !ok {
		return
	}
	if strings.HasPrefix(id, "config:") {
		writeErr(w, http.StatusBadRequest, "config-managed agents must be edited in config.toml")
		return
	}
	if _, authorized := s.authorizeAgent(w, r, id, core.GrantLevelManage); !authorized {
		return
	}
	if err := s.st.DeleteAgentInstance(r.Context(), id); err != nil {
		writeErr(w, http.StatusInternalServerError, err.Error())
		return
	}
	if s.connect != nil {
		if err := s.connect.RestartChannelsForAgent(r.Context(), id); err != nil {
			writeErr(w, http.StatusInternalServerError, "agent deleted, but bound channels failed to restart: "+err.Error())
			return
		}
	}
	writeOK(w)
}

func (s *Server) agentInstances(ctx context.Context) ([]core.AgentInstance, error) {
	return s.scopedAgentInstances(ctx, core.AdminPrincipal())
}

// scopedAgentInstances lists the Agent instances a principal may see. The
// admin additionally gets the synthetic config.toml agents, which are host
// infrastructure without an owner and therefore not shared with tenants.
func (s *Server) scopedAgentInstances(ctx context.Context, principal *core.Principal) ([]core.AgentInstance, error) {
	items := make([]core.AgentInstance, 0)
	if s.st != nil {
		var stored []core.AgentInstance
		var err error
		if principal.IsTenant() {
			stored, err = s.st.ListAgentInstancesForTenant(ctx, principal.TenantID)
		} else {
			stored, err = s.st.ListAgentInstances(ctx)
		}
		if err != nil {
			return nil, err
		}
		items = append(items, stored...)
	}
	if !principal.IsTenant() {
		seen := map[string]bool{}
		for _, item := range items {
			seen[item.ID] = true
		}
		for _, item := range s.configAgentInstances() {
			if !seen[item.ID] {
				items = append(items, item)
			}
		}
	}
	s.enrichAgentProviders(ctx, items, principal)
	s.labelAgentOwners(ctx, items)
	sort.SliceStable(items, func(i, j int) bool {
		if items[i].Enabled != items[j].Enabled {
			return items[i].Enabled
		}
		return strings.ToLower(items[i].Name) < strings.ToLower(items[j].Name)
	})
	return items, nil
}

func (s *Server) findAgentInstance(ctx context.Context, id string) (core.AgentInstance, bool, error) {
	items, err := s.agentInstances(ctx)
	if err != nil {
		return core.AgentInstance{}, false, err
	}
	for _, item := range items {
		if item.ID == id {
			return item, true, nil
		}
	}
	return core.AgentInstance{}, false, nil
}

func (s *Server) normalizeAgentInstance(ctx context.Context, a *core.AgentInstance) error {
	a.Name = strings.TrimSpace(a.Name)
	a.RuntimeID = strings.TrimSpace(a.RuntimeID)
	a.DesktopThreadID = strings.TrimSpace(a.DesktopThreadID)
	a.WorkDir = strings.TrimSpace(a.WorkDir)
	a.WorkspaceMode = strings.ToLower(strings.TrimSpace(a.WorkspaceMode))
	a.WorktreeBaseRef = strings.TrimSpace(a.WorktreeBaseRef)
	a.SessionBackend = strings.ToLower(strings.TrimSpace(a.SessionBackend))
	a.ProviderTool = strings.TrimSpace(a.ProviderTool)
	a.ProviderID = strings.TrimSpace(a.ProviderID)
	a.DefaultModel = strings.TrimSpace(a.DefaultModel)
	a.DefaultReasoningEffort = strings.TrimSpace(a.DefaultReasoningEffort)
	a.DefaultServiceTier = strings.TrimSpace(a.DefaultServiceTier)
	a.DefaultApprovalMode = strings.TrimSpace(a.DefaultApprovalMode)
	a.MemoryScope = strings.TrimSpace(a.MemoryScope)
	if a.Name == "" {
		return fmt.Errorf("agent name is required")
	}
	if a.RuntimeID == "" {
		return fmt.Errorf("runtime is required")
	}
	if !knownAgentRuntime(a.RuntimeID) {
		return fmt.Errorf("unknown agent runtime %q", a.RuntimeID)
	}
	if err := s.normalizeCodexDesktopAgent(ctx, a); err != nil {
		return err
	}
	if a.WorkspaceMode == "" {
		a.WorkspaceMode = "shared"
	}
	if a.WorkspaceMode != "shared" && a.WorkspaceMode != "worktree" {
		return fmt.Errorf("workspace mode must be shared or worktree")
	}
	if a.WorkspaceMode == "worktree" && a.WorkDir == "" {
		return fmt.Errorf("worktree workspace mode requires a working directory")
	}
	if a.SessionBackend == "" {
		a.SessionBackend = "structured"
	}
	if a.SessionBackend != "structured" && a.SessionBackend != "tmux" {
		return fmt.Errorf("session backend must be structured or tmux")
	}
	if a.SessionBackend == "tmux" {
		spec, ok := framework.Lookup(a.RuntimeID)
		if !ok || spec.KindType != framework.KindCLI {
			return fmt.Errorf("tmux session backend requires a CLI runtime")
		}
	}
	newRecord := strings.TrimSpace(a.ID) == ""
	creatingRecord := newRecord
	if !newRecord && s.st != nil {
		existing, err := s.st.GetAgentInstance(ctx, a.ID)
		if err != nil {
			return fmt.Errorf("load existing agent: %w", err)
		}
		creatingRecord = existing == nil
	}
	if creatingRecord && !agentRuntimeAvailable(a.RuntimeID) {
		return fmt.Errorf("agent runtime %q is not installed on this machine", a.RuntimeID)
	}
	if a.DefaultApprovalMode != "" && !core.ApprovalModeSupported(a.RuntimeID, a.DefaultApprovalMode) {
		return fmt.Errorf("approval mode %q is not supported by %s", a.DefaultApprovalMode, a.RuntimeID)
	}
	if newRecord {
		a.ID = "agent-" + randHex(6)
	}
	if strings.HasPrefix(a.ID, "config:") {
		return fmt.Errorf("config-managed agent ids are read-only")
	}
	now := time.Now()
	if !newRecord && a.CreatedAt.IsZero() && s.st != nil {
		if existing, err := s.st.GetAgentInstance(ctx, a.ID); err == nil && existing != nil {
			a.CreatedAt = existing.CreatedAt
		}
	}
	if a.CreatedAt.IsZero() {
		a.CreatedAt = now
	}
	a.UpdatedAt = now
	if a.ProviderTool == "" {
		a.ProviderTool = a.RuntimeID
	}
	if err := s.validateAgentDefaultRuntimeSettings(ctx, a); err != nil {
		return err
	}
	if a.MemoryScope == "" {
		a.MemoryScope = "agent:" + a.ID
	}
	if a.Source == "config.toml" {
		a.Source = "manual"
	}
	if a.Source == "" {
		if newRecord {
			a.Source = "manual"
		} else {
			a.Source = "console"
		}
	}
	if newRecord && len(a.ChannelBindings) == 0 {
		a.ChannelBindings = []core.AgentChannelBinding{}
	}
	return nil
}

func (s *Server) normalizeCodexDesktopAgent(ctx context.Context, a *core.AgentInstance) error {
	if a.RuntimeID != "codex-app" {
		a.DesktopThreadID = ""
		return nil
	}
	if a.DesktopThreadID == "" {
		return fmt.Errorf("Codex Desktop thread is required")
	}
	if s.sessions == nil {
		return fmt.Errorf("Codex Desktop session service is unavailable")
	}
	threads, err := s.sessions.List(ctx, "codex", "desktop")
	if err != nil {
		return fmt.Errorf("list Codex Desktop threads: %w", err)
	}
	var selected *sessionstore.Meta
	for i := range threads {
		if threads[i].SessionID == a.DesktopThreadID {
			selected = &threads[i]
			break
		}
	}
	if selected == nil {
		return fmt.Errorf("Codex Desktop thread %q was not found", a.DesktopThreadID)
	}
	if strings.TrimSpace(selected.ProjectDir) == "" {
		return fmt.Errorf("Codex Desktop thread %q has no working directory", a.DesktopThreadID)
	}
	if a.WorkDir != "" && !sameCanonicalPath(a.WorkDir, selected.ProjectDir) {
		return fmt.Errorf("Codex Desktop thread belongs to %q, agent uses %q", selected.ProjectDir, a.WorkDir)
	}
	a.WorkDir = selected.ProjectDir
	a.WorkspaceMode = "shared"
	a.WorktreeBaseRef = ""
	a.SessionBackend = "structured"
	if s.st == nil {
		return nil
	}
	existing, err := s.st.ListAgentInstances(ctx)
	if err != nil {
		return fmt.Errorf("check Codex Desktop thread binding: %w", err)
	}
	for _, item := range existing {
		if item.ID != a.ID && item.RuntimeID == "codex-app" && item.DesktopThreadID == a.DesktopThreadID {
			return fmt.Errorf("Codex Desktop thread is already bound to agent %q", item.Name)
		}
	}
	return nil
}

func (s *Server) configAgentInstances() []core.AgentInstance {
	if s.cfg == nil {
		return nil
	}
	out := make([]core.AgentInstance, 0, len(s.cfg.Projects))
	for projectIndex, p := range s.cfg.Projects {
		id := configAgentInstanceID(p.Name, projectIndex)
		channels := make([]core.AgentChannelBinding, 0, len(p.Platforms))
		for i, raw := range p.Platforms {
			typ, _ := raw["type"].(string)
			if typ == "" {
				typ = "platform"
			}
			channels = append(channels, core.AgentChannelBinding{
				ID:     id + ":channel:" + strconv.Itoa(i),
				Type:   typ,
				Name:   typ,
				Status: "configured",
				Config: redactAnyMap(raw),
			})
		}
		out = append(out, core.AgentInstance{
			ID:              id,
			Name:            p.Name,
			RuntimeID:       p.Agent,
			WorkDir:         p.WorkDir,
			WorkspaceMode:   firstAgentWorkspaceMode(p.WorkspaceMode),
			WorktreeBaseRef: p.WorktreeBaseRef,
			SessionBackend:  firstAgentSessionBackend(p.SessionBackend),
			SystemPrompt:    p.SystemPrompt,
			ProviderTool:    p.Agent,
			DefaultModel:    p.DefaultModel,
			MemoryScope:     "project:" + p.Name,
			Env:             redactStringMap(p.Env),
			ChannelBindings: channels,
			Enabled:         true,
			Source:          "config.toml",
		})
	}
	return out
}

func firstAgentWorkspaceMode(value string) string {
	if strings.EqualFold(strings.TrimSpace(value), "worktree") {
		return "worktree"
	}
	return "shared"
}

func firstAgentSessionBackend(value string) string {
	if strings.EqualFold(strings.TrimSpace(value), "tmux") {
		return "tmux"
	}
	return "structured"
}

func configAgentInstanceID(projectName string, projectIndex int) string {
	return "config:" + slugID(projectName) + "-" + strconv.Itoa(projectIndex+1)
}

func (s *Server) validateAgentDefaultRuntimeSettings(ctx context.Context, a *core.AgentInstance) error {
	if a.DefaultModel == "" && a.DefaultReasoningEffort == "" && a.DefaultServiceTier == "" && a.DefaultApprovalMode == "" {
		return nil
	}
	p, err := s.agentProvider(ctx, a)
	if err != nil {
		return err
	}
	if p == nil {
		// Local Codex login discovers its catalog only after app-server starts.
		// Let that runtime validate the values instead of rejecting the Agent
		// record before it can reach the signed-in account.
		return nil
	}
	if a.DefaultModel != "" {
		models := core.ProviderModelOptions(p)
		if len(models) == 0 {
			return fmt.Errorf("provider %q has no selectable models", p.ID)
		}
		if !containsRuntimeValue(models, a.DefaultModel) {
			return fmt.Errorf("default model %q is not supported by provider %q", a.DefaultModel, p.ID)
		}
	}
	if a.DefaultReasoningEffort != "" && len(p.Meta.SupportedReasoningEfforts) > 0 &&
		!containsRuntimeValue(p.Meta.SupportedReasoningEfforts, a.DefaultReasoningEffort) {
		return fmt.Errorf("default reasoning effort %q is not supported by provider %q", a.DefaultReasoningEffort, p.ID)
	}
	if a.DefaultServiceTier != "" && len(p.Meta.SupportedServiceTiers) > 0 &&
		!containsRuntimeValue(p.Meta.SupportedServiceTiers, a.DefaultServiceTier) {
		return fmt.Errorf("default service tier %q is not supported by provider %q", a.DefaultServiceTier, p.ID)
	}
	return nil
}

func containsRuntimeValue(values []string, target string) bool {
	for _, value := range values {
		if value == target {
			return true
		}
	}
	return false
}

func (s *Server) agentProvider(ctx context.Context, a *core.AgentInstance) (*core.Provider, error) {
	if s.provider == nil {
		return nil, nil
	}
	if a.ProviderID != "" {
		return s.provider.Get(ctx, a.ProviderID)
	}
	tool := a.ProviderTool
	if tool == "" {
		tool = a.RuntimeID
	}
	routes, err := s.provider.ActiveRoutes(ctx)
	if err != nil {
		return nil, err
	}
	want := core.NormalizeProviderTool(tool)
	for _, route := range routes {
		if route.Tool == tool {
			return s.provider.Get(ctx, route.ProviderID)
		}
	}
	for _, route := range routes {
		if core.NormalizeProviderTool(route.Tool) == want {
			return s.provider.Get(ctx, route.ProviderID)
		}
	}
	return nil, nil
}

func (s *Server) enrichAgentProviders(ctx context.Context, items []core.AgentInstance, principal *core.Principal) {
	if s.provider == nil {
		return
	}
	var providers []*core.Provider
	var err error
	if principal.IsTenant() && s.st != nil {
		providers, err = s.st.ListProvidersForTenant(ctx, principal.TenantID)
	} else {
		providers, err = s.provider.List(ctx)
	}
	if err != nil {
		return
	}
	names := map[string]string{}
	for _, p := range providers {
		names[p.ID] = p.Name
	}
	for i := range items {
		if items[i].ProviderID != "" {
			items[i].ProviderName = names[items[i].ProviderID]
		}
	}
}

func knownAgentRuntime(id string) bool {
	for _, name := range core.RegisteredAgents() {
		if name == id {
			return true
		}
	}
	return false
}

func availableAgentRuntimes() []string {
	registered := core.RegisteredAgents()
	available := make([]string, 0, len(registered))
	for _, id := range registered {
		if agentRuntimeAvailable(id) {
			available = append(available, id)
		}
	}
	return available
}

func agentRuntimeAvailable(id string) bool {
	if id == "codex-app" {
		return framework.IsInstalled("codex")
	}
	if spec, catalogued := framework.Lookup(id); catalogued {
		if spec.Hidden {
			return false
		}
		return framework.IsInstalled(id)
	}
	// Third-party adapters do not have a built-in framework specification. A
	// successful registration is their declaration that they are runnable.
	return core.HasAgent(id)
}

func randHex(n int) string {
	b := make([]byte, n)
	if _, err := rand.Read(b); err != nil {
		return strconv.FormatInt(time.Now().UnixNano(), 36)
	}
	return hex.EncodeToString(b)
}

func slugID(s string) string {
	s = strings.ToLower(strings.TrimSpace(s))
	var b strings.Builder
	lastDash := false
	for _, r := range s {
		ok := r >= 'a' && r <= 'z' || r >= '0' && r <= '9'
		if ok {
			b.WriteRune(r)
			lastDash = false
			continue
		}
		if !lastDash {
			b.WriteByte('-')
			lastDash = true
		}
	}
	out := strings.Trim(b.String(), "-")
	if out == "" {
		return "agent"
	}
	return out
}

func redactStringMap(in map[string]string) map[string]string {
	if len(in) == 0 {
		return nil
	}
	out := map[string]string{}
	for k, v := range in {
		if isSecretish(k) {
			out[k] = "<redacted>"
		} else {
			out[k] = v
		}
	}
	return out
}

func redactAnyMap(in map[string]any) map[string]string {
	if len(in) == 0 {
		return nil
	}
	out := map[string]string{}
	for k, v := range in {
		if isSecretish(k) {
			out[k] = "<redacted>"
			continue
		}
		switch x := v.(type) {
		case string:
			out[k] = x
		case fmt.Stringer:
			out[k] = x.String()
		default:
			b, err := json.Marshal(x)
			if err == nil {
				out[k] = string(b)
			}
		}
	}
	return out
}

func isSecretish(key string) bool {
	key = strings.ToLower(key)
	for _, needle := range []string{"secret", "token", "password", "api_key", "apikey", "app_secret", "key"} {
		if strings.Contains(key, needle) {
			return true
		}
	}
	return false
}